+++ app/src/main/java/com/example/movietop/domain/movies/list/Film.kt
@@ -1,4 +1,4 @@
-package com.example.movietop.domain.movies
+package com.example.movietop.domain.movies.list
 
 data class Film(
     val id: Int,
