+++ app/src/main/java/com/example/movietop/domain/movies/list/MoviesListRepository.kt
@@ -1,7 +1,6 @@
 package com.example.movietop.domain.movies.list
 
 import androidx.paging.PagingData
-import com.example.movietop.domain.movies.Film
 import kotlinx.coroutines.flow.Flow
 
 interface MoviesListRepository {
