+++ app/src/main/java/com/example/movietop/ui/movies/list/MoviesListViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
-import com.example.movietop.domain.movies.Film
+import com.example.movietop.domain.movies.list.Film
 import com.example.movietop.domain.movies.list.MoviesListRepository
 import kotlinx.coroutines.flow.Flow
