+++ app/src/main/java/com/example/movietop/utils/Mapper.kt
@@ -1,7 +1,7 @@
 package com.example.movietop.utils
 
 import com.example.movietop.data.movies.list.FilmDTO
-import com.example.movietop.domain.movies.Film
+import com.example.movietop.domain.movies.list.Film
 
 fun FilmDTO.toFilm() = Film(
     id = id,
